+++ code-types/TypeConverter.ts
@@ -117,7 +117,7 @@ module TypeConverter {
 
 
     /** Check if a type is a core built-in C#, Java, or TypeScript data type.
-     * A core type is a widely support type that is core to the language - (i.e. C# has string and DateTime)
+     * A core type is a widely support type that is core to the language but not a primitive - (i.e. C# has string and DateTime)
      * @param typeName the simple data type name (i.e. 'number' or 'byte')
      */
     export function isCore(typeName: string): boolean {
@@ -140,6 +140,22 @@ module TypeConverter {
     export function isGeneric(dataType: CodeAst.Type) {
         return dataType && dataType.genericParameters != null && dataType.genericParameters.length > 0;
     }
+
+
+    /** Check whether an object is an Open API Spec type
+     * @param obj the object to test
+     */
+    export function isOpenApiType(obj: any): obj is OpenApiV2.Type {
+        return "type" in obj && (obj["type"] === "array" ? "items" in obj : (obj["type"] === "object" ? "properties" in obj : true));
+    }
+
+
+	/** Check whether an object is an Open API Reference type
+	 * @param obj the object to test
+	 */
+	export function isOpenApiReference(obj: any): obj is OpenApiV3.Reference {
+		return "$ref" in obj;
+	}
 
 
     /** Create a deep copy of a CodeAst.Type
@@ -170,6 +186,7 @@ module TypeConverter {
             arrayDimensions: type.arrayDimensions,
             genericParameters: resParams,
             nullable: type.nullable,
+            primitive: type.primitive,
             typeName: type.typeName,
         };
 
@@ -184,7 +201,22 @@ module TypeConverter {
 
     /** Functions for transforming and manipulating type information into TypesScript types (i.e. number, boolean, string)
      */
-    export class TypeScript {
+    export module TypeScript {
+
+        /** {@link Date} instance function names that return a string (or number) representation of the date */
+        export type DateGetter =
+            "getTime" |
+            "toDateString" |
+            "toISOString" |
+            "toJSON" |
+            "toLocaleString" |
+            "toLocaleDateString" |
+            "toLocaleTimeString" |
+            "toString" |
+            "toTimeString" |
+            "toUTCString" |
+            "valueOf";
+
 
         /** Given a string or a Type, convert either to a TypeScript type:
          * If string, assume it is a simple typeTemplate (see parseAndConvertTypeTemplate()), parse it and convert the types to TypeScript.
@@ -192,7 +224,7 @@ module TypeConverter {
          * @param typeTemplate a Type, or string where the format must be 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
          * @param returnUnknownTypes
          */
-        static parseTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean): CodeAst.Type {
+        export function parseTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean): CodeAst.Type {
             if (typeof typeTemplate === "string") {
                 var parsedType = TypeConverter.parseTypeTemplate(typeTemplate);
                 parsedType.typeName = TypeScript.convertSimpleType(parsedType.typeName, returnUnknownTypes);
@@ -214,7 +246,7 @@ module TypeConverter {
          * @param includeNullability
          * @param [nullableSymbol="?"]
          */
-        static parseAndConvertTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean, includeNullability: boolean, nullableSymbol = "?"): string {
+        export function parseAndConvertTypeTemplate(typeTemplate: string | CodeAst.Type, returnUnknownTypes: boolean, includeNullability: boolean, nullableSymbol = "?"): string {
             return (typeof typeTemplate === "string"
                 ? TypeScript.parseAndConvertTypeTemplateString(typeTemplate, returnUnknownTypes, includeNullability, nullableSymbol)
                 : typeToString(typeTemplate, (t) => TypeScript.convertSimpleType(t, returnUnknownTypes), includeNullability, nullableSymbol));
@@ -228,7 +260,7 @@ module TypeConverter {
          * @param includeNullability
          * @param nullableSymbol the symbol to put after nullable types (i.e. '?' for C# or ' | null' for TypeScript)
          */
-        static parseAndConvertTypeTemplateString(typeTemplate: string, returnUnknownTypes: boolean, includeNullability: boolean, nullableSymbol: string): string {
+        export function parseAndConvertTypeTemplateString(typeTemplate: string, returnUnknownTypes: boolean, includeNullability: boolean, nullableSymbol: string): string {
             var typeInfo = TypeConverter.parseTypeTemplate(typeTemplate);
             var arrayCount = typeInfo.arrayDimensions;
 
@@ -244,9 +276,9 @@ module TypeConverter {
 
         /** Convert primitive and common builtin C# and Java types to TypeScript equivalent types
          * @param typeName the type name (i.e. 'bool' or 'String')
-         * @param returnUnknownTypes
+         * @param returnUnknownTypes whether to return unknown 'typeName' values or throw an error
          */
-        static convertSimpleType(typeName: string, returnUnknownTypes: boolean): string {
+        export function convertSimpleType(typeName: string, returnUnknownTypes: boolean): string {
             switch (typeName) {
                 case "boolean":
                 case "string":
@@ -286,10 +318,12 @@ module TypeConverter {
 
 
         /** Create a TypeScript source code string that transforms a specific data type into a string (i.e. stringifies the type)
-         * @param typeTemplate
-         * @param variableName
+         * @param typeTemplate a CodeAst.Type or a type template string like 'typeName?[][]...' where typeName has no generic parameters, and the '?' (nullability) and '[][]...' (array dimensions) are optional
+         * @param variableName the name of the variable
+         * @param [dateToString] optional, function that returns source code that converts a given date variable to the format desired, if not provided 'toString()' is used for date/time related types
+         * @returns a line of source code that converts the variable from it's given type specified by 'typeTemplate' to a string
          */
-        static createTypeTemplateToStringCode(typeTemplate: string | CodeAst.Type, variableName: string): string {
+        export function createTypeTemplateToStringCode(typeTemplate: string | CodeAst.Type, variableName: string, dateToString?: DateGetter | ((variableName: string) => string)): string {
             var typeInfo = typeof typeTemplate === "string" ? TypeConverter.parseTypeTemplate(typeTemplate) : typeTemplate;
 
             if (typeInfo.genericParameters != null && typeInfo.genericParameters.length > 0) {
@@ -302,7 +336,14 @@ module TypeConverter {
             switch (typeName) {
                 case "bool":
                 case "boolean":
-                    if (<number>arrayCount > 0) { throw new Error("converting array of " + typeName + " to string not supported"); }
+                    if (arrayCount != null) {
+                        if (arrayCount > 1) {
+                            throw new Error("no source code to string conversion format for " + typeName + "[]".repeat(arrayCount) + "'");
+                        }
+                        else {
+                            return "(" + variableName + " ? Array.prototype.map.call(" + variableName + ", function (v) { return (v ? \"true\" : \"false\"); }) : \"null\")"
+                        }
+                    }
                     return "(" + variableName + " ? \"true\" : \"false\")";
                 case "byte":
                 case "sbyte":
@@ -318,10 +359,27 @@ module TypeConverter {
                 case "decimal":
                 case "real":
                 case "number":
+                    if (arrayCount != null && arrayCount > 1) {
+                        throw new Error("no source code to string conversion format for " + typeName + "[]".repeat(arrayCount) + "'");
+                    }
+                    return "(" + variableName + " ? " + variableName + ".toString() : \"null\")";
                 case "date":
                 case "DateTime":
+                    if (dateToString != null) {
+                        if (arrayCount != null) {
+                            if (arrayCount > 1) {
+                                throw new Error("no source code to string conversion format for '" + typeName + "[]".repeat(arrayCount) + "'");
+                            }
+                            else {
+                                return "(" + variableName + " ? Array.prototype.map.call(" + variableName + ", function (v) { return (v ? " + (typeof dateToString === "string" ? "v." + dateToString + "()" : dateToString("v")) + " : \"null\"); }) : \"null\")"
+                            }
+                        }
+                        return "(" + variableName + " ? " + (typeof dateToString === "string" ? variableName + "." + dateToString + "()" : dateToString(variableName)) + " : \"null\")";
+                    }
                 case "any":
-                    if (<number>arrayCount > 0) { throw new Error("converting array of " + typeName + " to strings not supported"); }
+                    if (arrayCount != null && arrayCount > 1) {
+                        throw new Error("no source code to string conversion format for " + typeName + "[]".repeat(arrayCount) + "'");
+                    }
                     return "(" + variableName + " ? " + variableName + ".toString() : \"null\")";
                 case "string":
                     return variableName;
@@ -331,12 +389,66 @@ module TypeConverter {
         }
 
 
-        /** Convert a map of property names and type template strings into an array of TypeScript source code strings that convert each property to a string */
-        static createTypeTemplatesToStringCode(props: StringMap<string>): string[] {
-            var keys = Object.keys(props);
+        /** Convert a map of property names and type templates into an array of TypeScript source code strings that convert each property to a string.
+         * @param props maps object property names to their property type templates
+         * @params [keys] optional, list of 'props' keys to map, if not provided 'Object.keys()' is used
+         * @returns an array of results from calling 'createTypeTemplateToStringCode()' on each property in the 'props' map
+         */
+        export function createTypeTemplatesToStringCode(props: StringMap<string | CodeAst.Type>, keys?: string[] | null): string[] {
+            keys = keys || Object.keys(props);
             return keys.map(k => TypeScript.createTypeTemplateToStringCode(props[k], k));
         }
 
+
+        /** Create a function which returns the source code string that converts a 'Date' variable into a string (or other type) when given a variable name.
+         * Can be used as the 3rd argument ('dateToString') when calling 'createTypeTemplateToStringCode()'
+         * @param dateToString the name of the 'Date' function to call on the variable to convert it
+         * @param [locale] optional, locale argument to pass to the 'Date' function defined by 'dateFunction'
+         * @param [dateTimeFormatOptions] optional, date-time format options to pass to the 'Date' function defined by 'dateFunction'
+         */
+        export function createDateToStringCodeFunction(dateToString: DateGetter, locale?: string | string[], dateTimeFormatOptions?: Intl.DateTimeFormatOptions) {
+            var args = (locale != null || dateTimeFormatOptions != null
+                ? (locale != null ? (typeof locale === "string" ? "\"" + locale + "\"" : "[\"" + locale.join("\",\"") + "\"]") : "undefined") +
+                    (dateTimeFormatOptions != null ? ", " + JSON.stringify(dateTimeFormatOptions) : "")
+                : "");
+
+            var dateStringify = dateToString + "(" + args + ")";
+
+            return function (variableName: string) {
+                return variableName + "." + dateStringify;
+            };
+        }
+    }
+
+
+
+    export module CSharp {
+
+        /** Convert an Open API Spec type to C# equivalent type.
+         * based on: https://swagger.io/specification/#data-types
+         * @param typeName the type name (i.e. 'bool' or 'String')
+         * @param format the 'format' field (if available) for this 'typeName' in the Open API document
+         * @param returnUnknownTypes whether to return unknown 'typeName' values or throw an error
+        */
+        export function convertOpenApiType(typeName: string, format: string | null | undefined, returnUnknownTypes: boolean): string {
+            switch (typeName) {
+                case "boolean":
+                    return "bool";
+                case "integer":
+                    return format === "int64" ? "long" : "int";
+                case "number":
+                    return format === "double" || format === "float" ? format : "int";
+                case "string":
+                    return format === "date" || format === "date-time" ? "DateTime" : "string";
+                default:
+                    if (returnUnknownTypes) {
+                        return typeName;
+                    }
+                    else {
+                        throw new Error("unknown type name for TypeScript code: '" + typeName + "'");
+                    }
+            }
+        }
     }
 
 }